06-kafka-seckill-decoupling/app/requirements.txt
flask==3.1.0
kafka-python==2.1.5
redis==5.2.1
